Updated to support VTK 9.0 (#84)
* VTK_USE_FILE -> vtk_module_autoinit

* Avoid use of deprecated GetCellPoints(). Avoid int->size_t warning.

* Updated VTK modules list to cope with version 9.0

* Reactivated builds for testing

* If VTK < 7 then need vtkRenderingOpenGL instead of vtkRenderingOpenGL2

* Added Ubuntu 20.04 and MacOS 10.15.7 platforms for testing.

* Removed stray else()

* Disabled 20.04 clang build pending a fix for __exp_finite. Added qt5 to MacOS 10.15.7 build.

* Need qt5 for MacOS 10.13 build too?

* Disabled MacOS 10.13 and 10.15.7 builds pending a fix for a qt5 error in cmake

+++ src/readybase/ImageRD.hpp
@@ -79,7 +79,7 @@ class ImageRD : public AbstractRD
         virtual void SetValue(float x,float y,float z,float val,const Properties& render_settings);
         virtual void SetValuesInRadius(float x,float y,float z,float r,float val,const Properties& render_settings);
 
-        virtual int GetMemorySize() const;
+        virtual size_t GetMemorySize() const;
 
     protected:
